optimize connector struct & streamx-flink-connector support scala and java api (#892)
* optimize connector struct && adapt java and scala api
