+++ streamx-flink/streamx-flink-connector/streamx-flink-connector-clickhouse/src/main/scala/com/streamxhub/streamx/flink/connector/clickhouse/scala/util/PoJoConvertUtils.scala
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.streamxhub.streamx.flink.connector.clickhouse.scala.util
+
+object PoJoConvertUtils {
+
+  def parsePojoToInsertValue[T](value: T): String = {
+    val buffer = new StringBuilder("(")
+    val fields = value.getClass.getDeclaredFields
+    fields.foreach(f => {
+      f.setAccessible(true)
+      val v = f.get(value)
+      f.getType.getSimpleName match {
+        case "String" => buffer.append(s""""$v",""".stripMargin)
+        case _ => buffer.append(s"""$v,""".stripMargin)
+      }
+    })
+    buffer.toString().replaceFirst(",$", ")")
+  }
+}
