+++ streamx-flink/streamx-flink-connector/streamx-flink-connector-influxdb/src/main/scala/com/streamxhub/streamx/flink/connector/influxdb/scala/sink/InfluxDBOutputFormat.scala
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.streamxhub.streamx.flink.connector.influxdb.scala.sink
+
+import com.streamxhub.streamx.common.util.Logger
+import com.streamxhub.streamx.flink.connector.influxdb.scala.domian.InfluxEntity
+import com.streamxhub.streamx.flink.connector.influxdb.scala.internal.InfluxDBFunction
+import org.apache.flink.api.common.io.RichOutputFormat
+import org.apache.flink.api.common.typeinfo.TypeInformation
+import org.apache.flink.configuration.Configuration
+
+import java.util.Properties
+
+class InfluxDBOutputFormat[T: TypeInformation](implicit prop: Properties, endpoint: InfluxEntity[T]) extends RichOutputFormat[T] with Logger {
+
+  private val sinkFunction = new InfluxDBFunction[T](prop)
+
+  private var configuration: Configuration = _
+
+  override def configure(configuration: Configuration): Unit = this.configuration = configuration
+
+  override def open(taskNumber: Int, numTasks: Int): Unit = sinkFunction.open(this.configuration)
+
+  override def writeRecord(record: T): Unit = sinkFunction.invoke(record, null)
+
+  override def close(): Unit = sinkFunction.close()
+}
