+++ streamx-flink/streamx-flink-connector/streamx-flink-connector-redis/src/main/scala/com/streamxhub/streamx/flink/connector/redis/scala/internal/RedisTransaction.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.streamxhub.streamx.flink.connector.redis.scala.internal
+
+import com.streamxhub.streamx.common.util.Utils
+import com.streamxhub.streamx.flink.connector.redis.scala.domain.RedisMapper
+
+import scala.collection.mutable
+
+case class RedisTransaction[T](
+                                transactionId: String = Utils.uuid(),
+                                mapper: mutable.MutableList[(RedisMapper[T], T, Int)] = mutable.MutableList.empty[(RedisMapper[T], T, Int)],
+                                var invoked: Boolean = false) extends Serializable {
+  def +(redisMapper: (RedisMapper[T], T, Int)): Unit = mapper += redisMapper
+
+  override def toString: String = s"(transactionId:$transactionId,size:${mapper.size},invoked:$invoked)"
+}
